+++ src/neuroconv/tools/nwb_helpers/__init__.py
@@ -1,26 +1,19 @@
 """Collection of Pydantic models and helper functions for configuring dataset IO parameters for different backends."""
-from ._dataset_configuration import (
-    configure_backend,
-    get_default_backend_configuration,
-    get_default_dataset_io_configurations,
-)
+from ._dataset_configuration import get_default_dataset_io_configurations
+from ._backend_configuration import get_default_backend_configuration
+from ._configure_backend import configure_backend
+from ._configuration_models._base_backend import BackendConfiguration
+from ._configuration_models._base_dataset_io import DatasetInfo, DatasetIOConfiguration
+from ._configuration_models._hdf5_backend import HDF5BackendConfiguration
+from ._configuration_models._hdf5_dataset_io import AVAILABLE_HDF5_COMPRESSION_METHODS, HDF5DatasetIOConfiguration
+from ._configuration_models._zarr_backend import ZarrBackendConfiguration
+from ._configuration_models._zarr_dataset_io import AVAILABLE_ZARR_COMPRESSION_METHODS, ZarrDatasetIOConfiguration
 from ._metadata_and_file_helpers import (
     add_device_from_metadata,
     get_default_nwbfile_metadata,
     get_module,
     make_nwbfile_from_metadata,
     make_or_load_nwbfile,
-)
-from ._models._base_models import DatasetInfo
-from ._models._hdf5_models import (
-    AVAILABLE_HDF5_COMPRESSION_METHODS,
-    HDF5BackendConfiguration,
-    HDF5DatasetIOConfiguration,
-)
-from ._models._zarr_models import (
-    AVAILABLE_ZARR_COMPRESSION_METHODS,
-    ZarrBackendConfiguration,
-    ZarrDatasetIOConfiguration,
 )
 
 BACKEND_CONFIGURATIONS = dict(hdf5=HDF5BackendConfiguration, zarr=ZarrBackendConfiguration)
@@ -30,6 +23,12 @@ __all__ = [
     "get_default_backend_configuration",
     "get_default_dataset_io_configurations",
     "configure_backend",
+    "AVAILABLE_HDF5_COMPRESSION_METHODS",
+    "AVAILABLE_ZARR_COMPRESSION_METHODS",
+    "BackendConfiguration",
+    "DatasetIOConfiguration",
+    "get_default_dataset_io_configurations",
+    "get_default_backend_configuration",
     "add_device_from_metadata",
     "get_default_nwbfile_metadata",
     "get_module",